requirements.txt
Flask
Flask-SQLAlchemy
Flask-migrate
python-dotenv
prometheus_flask_exporter
prometheus_client